libs/ai_service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paint-ai-service"
version = "0.1.0"
description = "AI service for paint recommendations using LangChain"
authors = [{name = "mrcndz"}]
dependencies = [
    "enrichment==0.1.0",
    "fastapi==0.115.13",
    "httpx==0.28.1",
    "langchain==0.3.25",
    "langchain_core==0.3.65",
    "langchain_openai==0.3.24",
    "pgvector==0.4.1",
    "pydantic==2.11.7",
    "pytest==8.4.1",
    "python-dotenv==1.1.0",
    "SQLAlchemy==2.0.23",
    "uvicorn==0.34.3",
    "asyncpg",
    "psycopg2",
    "PyJWT>=2.8.0",
    "shared",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=4.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["*"]
